seerah-assistant/requirements.txt
langchain
langchain-community
langchain-openai
pypdf
faiss-cpu
python-dotenv
tiktoken
sentence-transformers